5.9提交

+++ src/main/java/com/hawksoft/platform/controller/FlowController.java
@@ -4,6 +4,7 @@ import com.alibaba.fastjson.JSON;
 
 import com.hawksoft.platform.entity.Flow;
 
+import com.hawksoft.platform.entity.SpeedFlow;
 import com.hawksoft.platform.entity.Water;
 import com.hawksoft.platform.service.FlowService;
 
@@ -11,6 +12,7 @@ import com.hawksoft.platform.service.SpeedFlowService;
 import com.hawksoft.platform.service.VideoService;
 import com.hawksoft.platform.service.WaterStationService;
 import com.hawksoft.platform.socket.SocketUtils;
+import com.hawksoft.platform.util.DataUtil;
 import com.hawksoft.platform.util.DateUtil;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -152,6 +154,10 @@ public class FlowController {
     @RequestMapping("/saveFlow")
     @ResponseBody
     public String saveFlow(Flow flow){
+        if (flow.getCollectionTime().isEmpty()) {
+            logger.error("DATE format ERROR");
+            return "fail";
+        }
         return (flowService.saveFlow(flow) != 0) ? "success" : "fail";
     }
 
@@ -162,6 +168,11 @@ public class FlowController {
     @RequestMapping("/updateFlow")
     @ResponseBody
     public String updateFlow(Flow flow){
+        Object[] objects={flow,flow.getFlowId(),flow.getCollectionTime()};
+        if (!DataUtil.anyNotEmpty(objects)) {
+            logger.error("Parameter ERROR");
+            return "fail";
+        }
         return (flowService.updateFlow(flow) != 0) ? "success" : "fail";
     }
 
@@ -172,6 +183,11 @@ public class FlowController {
     @RequestMapping("/deleteFlow")
     @ResponseBody
     public String deleteFlow(Flow flow){
+        Object[] objects={flow,flow.getFlowId(),flow.getCollectionTime()};
+        if (!DataUtil.anyNotEmpty(objects)) {
+            logger.error("Parameter ERROR");
+            return "fail";
+        }
         return (flowService.deleteFlow(flow)  != 0) ? "success" : "fail";
     }
 
@@ -285,64 +301,6 @@ public class FlowController {
         return res;
     }
 
-    /**
-     * 根据站点ID获取相机拍摄水位信息
-     */
-    @RequestMapping(value = "/RealTimeAcquisitionData/{stnId}", method = RequestMethod.GET)
-    @ResponseBody
-    public String RealTimeAcquisitionData(@PathVariable("stnId") int stnId) {
-        String realTimeAcquisitionData="";
-        //通过站点ID获取站点名称
-        String stnCode = waterStationService.queryCodeById(stnId);
-        if (stnCode.isEmpty()) {
-            return "该站点不支持实时采集";
-        }
-        SimpleDateFormat simpleDateFormat=new SimpleDateFormat("yyyy-MM-dd HH:mm:ss");
-        //获取发送请求时间
-        String sendTime=DateUtil.getNowDate();
-        logger.debug("startTime:"+sendTime);
-
-        Socket so = SocketUtils.findSocket(stnCode);
-        if (so != null && so.isConnected()) {
-            // build message and send to RTU
-            StringBuilder message=new StringBuilder("CMD:1:FLUX");
-            try {
-                SocketUtils.sendMessage(so, message.toString());
-                int i=0;
-                do {
-                    i++;
-                    //获取发送完时间
-                    String receiveTime=DateUtil.getNowDate();
-                    logger.debug("endTime:"+receiveTime);
-
-                    Map<String,Object> timeMap=new HashMap<>();
-                    timeMap.put("startTime",sendTime);
-                    timeMap.put("endTime",receiveTime);
-                    timeMap.put("stnId",stnId);
-                    List<Flow> flowList=flowService.queryAllFlowByTime(timeMap);
-                    if (flowList.size()>0) {
-                        realTimeAcquisitionData= JSON.toJSON(flowList.get(0)).toString();
-                        break;
-                    }
-                    Thread.sleep(100);
-                } while (i<=1200);
-            } catch (IOException e) {
-                logger.error("Send scoket message error");
-                e.printStackTrace();
-            }
-            catch (InterruptedException e) {
-                logger.error("Thread running error");
-                e.printStackTrace();
-            }
-        } else {
-            return "The connection with RTU was lost!";
-        }
-
-        if(realTimeAcquisitionData.equals("")){
-            return "RealTimeAcquisitionData is failing";
-        }
-        return realTimeAcquisitionData;
-    }
 
     @RequestMapping(value = "/TestHighConcurrency/{stnId}", method = RequestMethod.GET)
     @ResponseBody
